Completed Marsh Area
Completed obstacles for marsh area.

Changed all trigger zones to have a width of 0.2

Removed function UpdateWindDuration() under movement script.

Fixed an issue where player could be stuck on the left side of the tree. (by adding an invisible block).

Work around issue of grapple block not being able to grapple at certain locations is to simply not have the player standing on a block in the trigger zone when grappling.